Made requested changes from pull request.
