+++ .idea/wcEcoli.iml
@@ -9,7 +9,7 @@
       <excludeFolder url="file://$MODULE_DIR$/runscripts/metabolism/homeostaticFBA_modularFBA_toy_model/svg_plots" />
       <excludeFolder url="file://$MODULE_DIR$/user" />
     </content>
-    <orderEntry type="jdk" jdkName="Python 3.8 (wcEcoli4)" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Python 3.8 (wcEcoli3)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" name="R User Library" level="project" />
     <orderEntry type="library" name="R Skeletons" level="application" />
